Create DHT Sensor
DHT11 Temperature & Humidity Sensor features a temperature & humidity sensor complex with a 
calibrated digital signal output. 
The DHT11 is a basic, ultra low-cost digital temperature and humidity sensor. It uses a capacitive 
humidity sensor and a thermistor to measure the surrounding air and spits out a digital signal on the 
data pin (no analog input pins are needed). It's fairly simple to use but requires careful timing to grab 
data